requirements.txt
pip
autopep8
Flask==3.0.3
requests
Werkzeug==3.0.6
imdbpy